// src/server/http/inovopay/inovopay.schema.ts
export namespace InovoPay {
// ...
  export const creditCardTransactionInputSchema = z.object({
    CUST_FNAME: z.string().nullish(),
    CUST_LNAME: z.string().nullish(),
    CUST_EMAIL: z.string().nullish(),
    LI_COUNT_1: z.number(),
    LI_PROD_ID_1: z.number(),
    LI_VALUE_1: z.number(),
    XTL_ORDER_ID: z.string().nullish(),
    BILL_ADDR: z.string().nullish(),
    BILL_ADDR_CITY: z.string().nullish(),
    BILL_ADDR_STATE: z.string().nullish(),
    BILL_ADDR_ZIP: z.string().nullish(),
    BILL_ADDR_COUNTRY: z.string().nullish(),
    PMT_NUMB: z.string(),
    TOKEN_GUID: z.string().nullish(),
    PMT_KEY: z.string(),
    PMT_EXPIRY: z.string(),
    CUST_LOGIN: z.string().nullish(),
    CUST_PASSWORD: z.string().nullish(),
    MERCH_ACCT_ID: z.string().nullish(),
    REQUEST_CURRENCY: z.string().default('USD'),
    PMT_DESCRIPTOR: z.string().nullish(),
    PMT_DESCRIPTOR_PHONE: z.string().nullish(),
    CUST_PHONE: z.string().nullish(),
    REQUEST_AFF_ID: z.string().nullish(),
    REQUEST_AFF_ID_SUB: z.string().nullish(),
    UNIQUE_XTL_ORDER_ID: z.enum(['0', '1', '2']).nullish(),
    PMT_ID_XTL: z.string().nullish(),
    MBSHIP_ID_XTL: z.string().nullish(),
    TRANS_REBILL_TYPE: z.enum(['NONE', 'TRIAL', 'INITIAL', 'REBILL']).default('NONE'),
    TRANS_TRIAL_REBILL_CUSTOMER_CONSENT: z.string().nullish(),
    TRANS_CUSTOMER_RECEIPT: z.string().nullish(),
    CARD_ON_FILE_FLAG: z.enum(['0', '1']),
  });

  export type CreditCardTransactionInput = z.infer<typeof creditCardTransactionInputSchema>;
// ...
}